Split SDX activity e2e tests and extract shared SDX helpers.
Move org/runtime-group activity into 09-org-activity.ts, keep catalog
tests (including dual-endpoint subsystem lifecycle) in 08-catalog-activity.ts,
and consolidate gateway key helpers into sdx-commands.ts.
