+++ e2e/cypress/support/sdx-commands.ts
@@ -95,3 +95,85 @@ export function createRuntimeGroup(
     }
   )
 }
+
+export const publicKeyPemA = `-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA7YUiYt5uUxVY6yOzwZil
+5JlFJlRLxXmG08w/uOMb18Tfwb+5UZ4zEAIgiAgq2Fq+GKyiXN/qId9mySAiANUE
+HjjpnpOpAmKU6RsP+Emw54Fco/RMqkHGl2syNCWpgs+yqZ6ZXbw6wn5OfkaL0hB9
+id7p8yX/mxqH96ycdA/e3sZQ53X41EXfZl29E654K+LeEtMa+Hy0hIRz+bDOyptM
+yEllT/YWhWqhYA/JX+2VklnQ3k82dvvFGMGIS1yYkQuAIEg07TTEHcVAn31eov6T
++KHEVt70CdzgR9MK25U7u8V9Kp0JaKbmfPraCvo/BKzo/nNJa8RfIZvPvp/hKiSy
+HwIDAQAB
+-----END PUBLIC KEY-----`
+
+export const publicKeyPemB = `-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAtCakFxv/0+vpsocaigFF
+68fa6N2avit6LLLvcOwl6k/J13T3/mP/ALpOKyBmk0O/duiSeeUfsqm/Qbs3ASwW
+YCvpK6WRyo0xAdo7W3MvwPhHAIa7glic/LNvxuUZdWsW30sKrl2oRrvxPxY2lBkD
+wBg2VVA6Dq2cyusfegGDFq2e+f9YTildBljPOBHugXG3a+A/7ZRgKpIu4eu9U+Pj
+M12XMokf2owgpkT/b49bUYL0bqB+vzc+pViajneoxlPngRUVyRoZRduP7yK1p+bV
+S2jKMiCW1pr8CU26fbma7xHNoLGimmenkAqRhXiONSxKnsmGgoZvaFQzpqGKGxWc
+dQIDAQAB
+-----END PUBLIC KEY-----`
+
+export function orgGatewayKeyName(org: { tags: string[] }): string {
+  const memberClass = org.tags[0].split(':')[1].toLowerCase()
+  const memberId = org.tags[1].split(':')[1].toLowerCase()
+  return `sdx.keys.${memberClass}.${memberId}.org:0`
+}
+
+export function subsystemGatewayKeyName(clientId: string): string {
+  return `sdx.keys.${clientId.toLowerCase()}.sys:0`
+}
+
+export function registerHostOrganization(name: string, memberId: string) {
+  cy.setRequestBody({
+    name,
+    title: name,
+    description: 'e2e host org for runtime group activity',
+    tags: ['member_class:MIN', `member_id:${memberId}`],
+    orgUnits: [],
+    extSource: 'ckan',
+    extRecordHash: '',
+  })
+  return cy.callAPI('ds/api/v3/organizations/ca.bc.gov', 'PUT')
+}
+
+export function registerOrgGateway(orgName: string) {
+  cy.setRequestBody({})
+  return cy.callAPI(`ds/api/sdx/v1/organizations/${orgName}/gateway`, 'PUT')
+}
+
+export function applyOrgPublicKeyPattern(
+  orgName: string,
+  publicKeyPem: string,
+  action: 'apply' | 'remove' = 'apply'
+) {
+  cy.setRequestBody({
+    pattern: 'sdx-keys.r1',
+    parameters: {
+      organization: orgName,
+      public_key_pem: publicKeyPem,
+    },
+  })
+  cy.setQueryString({ action, dryRun: 'false' })
+  return cy.callAPI(`ds/api/sdx/v1/organizations/${orgName}/pattern`, 'PUT')
+}
+
+export function applySubsystemPublicKeyPattern(
+  orgName: string,
+  clientId: string,
+  publicKeyPem: string,
+  action: 'apply' | 'remove' = 'apply'
+) {
+  cy.setRequestBody({
+    pattern: 'sdx-keys.r1',
+    parameters: {
+      organization: orgName,
+      client_id: clientId,
+      public_key_pem: publicKeyPem,
+    },
+  })
+  cy.setQueryString({ action, dryRun: 'false' })
+  return cy.callAPI(`ds/api/sdx/v1/organizations/${orgName}/pattern`, 'PUT')
+}
