+++ e2e/cypress/tests/21-sdx-api/v1/08-catalog-activity.ts
@@ -1,94 +1,20 @@
 import { v4 as uuidv4 } from 'uuid'
 import {
+  applyOrgPublicKeyPattern,
+  applySubsystemPublicKeyPattern,
   clientIdForSubsystem,
   createOASService,
   createSubsystem,
+  orgGatewayKeyName,
+  publicKeyPemA,
+  publicKeyPemB,
+  registerOrgGateway,
+  subsystemGatewayKeyName,
   uniqueSubsystemName,
 } from '../../../support/sdx-commands'
 
-function registerHostOrganization(name: string, memberId: string) {
-  cy.setRequestBody({
-    name,
-    title: name,
-    description: 'e2e host org for runtime group activity',
-    tags: ['member_class:MIN', `member_id:${memberId}`],
-    orgUnits: [],
-    extSource: 'ckan',
-    extRecordHash: '',
-  })
-  return cy.callAPI('ds/api/v3/organizations/ca.bc.gov', 'PUT')
-}
-
-describe('SDX Organization Activity', () => {
+describe('SDX Catalog Activity', () => {
   let workingData: any
-
-  const orgGatewayKeyName = (org: { tags: string[] }) => {
-    const memberClass = org.tags[0].split(':')[1].toLowerCase()
-    const memberId = org.tags[1].split(':')[1].toLowerCase()
-    return `sdx.keys.${memberClass}.${memberId}.org:0`
-  }
-
-  const registerOrgGateway = (orgName: string) => {
-    cy.setRequestBody({})
-    return cy.callAPI(`ds/api/sdx/v1/organizations/${orgName}/gateway`, 'PUT')
-  }
-
-  const applyOrgPublicKeyPattern = (
-    orgName: string,
-    publicKeyPem: string,
-    action: 'apply' | 'remove' = 'apply'
-  ) => {
-    cy.setRequestBody({
-      pattern: 'sdx-keys.r1',
-      parameters: {
-        organization: orgName,
-        public_key_pem: publicKeyPem,
-      },
-    })
-    cy.setQueryString({ action, dryRun: 'false' })
-    return cy.callAPI(`ds/api/sdx/v1/organizations/${orgName}/pattern`, 'PUT')
-  }
-
-  const subsystemGatewayKeyName = (clientId: string) =>
-    `sdx.keys.${clientId.toLowerCase()}.sys:0`
-
-  const applySubsystemPublicKeyPattern = (
-    orgName: string,
-    clientId: string,
-    publicKeyPem: string,
-    action: 'apply' | 'remove' = 'apply'
-  ) => {
-    cy.setRequestBody({
-      pattern: 'sdx-keys.r1',
-      parameters: {
-        organization: orgName,
-        client_id: clientId,
-        public_key_pem: publicKeyPem,
-      },
-    })
-    cy.setQueryString({ action, dryRun: 'false' })
-    return cy.callAPI(`ds/api/sdx/v1/organizations/${orgName}/pattern`, 'PUT')
-  }
-
-  const publicKeyPemA = `-----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA7YUiYt5uUxVY6yOzwZil
-5JlFJlRLxXmG08w/uOMb18Tfwb+5UZ4zEAIgiAgq2Fq+GKyiXN/qId9mySAiANUE
-HjjpnpOpAmKU6RsP+Emw54Fco/RMqkHGl2syNCWpgs+yqZ6ZXbw6wn5OfkaL0hB9
-id7p8yX/mxqH96ycdA/e3sZQ53X41EXfZl29E654K+LeEtMa+Hy0hIRz+bDOyptM
-yEllT/YWhWqhYA/JX+2VklnQ3k82dvvFGMGIS1yYkQuAIEg07TTEHcVAn31eov6T
-+KHEVt70CdzgR9MK25U7u8V9Kp0JaKbmfPraCvo/BKzo/nNJa8RfIZvPvp/hKiSy
-HwIDAQAB
------END PUBLIC KEY-----`
-
-  const publicKeyPemB = `-----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAtCakFxv/0+vpsocaigFF
-68fa6N2avit6LLLvcOwl6k/J13T3/mP/ALpOKyBmk0O/duiSeeUfsqm/Qbs3ASwW
-YCvpK6WRyo0xAdo7W3MvwPhHAIa7glic/LNvxuUZdWsW30sKrl2oRrvxPxY2lBkD
-wBg2VVA6Dq2cyusfegGDFq2e+f9YTildBljPOBHugXG3a+A/7ZRgKpIu4eu9U+Pj
-M12XMokf2owgpkT/b49bUYL0bqB+vzc+pViajneoxlPngRUVyRoZRduP7yK1p+bV
-S2jKMiCW1pr8CU26fbma7xHNoLGimmenkAqRhXiONSxKnsmGgoZvaFQzpqGKGxWc
-dQIDAQAB
------END PUBLIC KEY-----`
 
   before(() => {
     cy.buildOrgGatewayDatasetAndProduct().then((data) => {
@@ -764,195 +690,6 @@ dQIDAQAB
           })
         }
       )
-    })
-  })
-
-  describe('Runtime group lifecycle in organization activity', () => {
-    let org: any;
-    let runtimeGroupId: string;
-
-    beforeEach(() => {
-      org = workingData.org;
-      runtimeGroupId = workingData.runtimeGroupId;
-      cy.setQueryString({});
-    });
-
-    it('records create on organization activity', () => {
-      const runtimeGroupName = `${runtimeGroupId}c`;
-      cy.setRequestBody({
-        name: runtimeGroupName,
-        hostedOrganizations: [org.name],
-      });
-      cy.callAPI(
-        `ds/api/sdx/v1/organizations/${org.name}/runtime-groups`,
-        'PUT'
-      ).then(({ apiRes: { status } }: any) => {
-        expect(status).to.be.equal(200);
-
-        cy.callAPI(
-          `ds/api/sdx/v1/organizations/${org.name}/activity?first=100`,
-          'GET'
-        ).then(({ apiRes: { status, body: activities } }: any) => {
-          expect(status).to.be.equal(200);
-          const created = activities.find(
-            (a: any) =>
-              a.params?.entity === 'RuntimeGroup' &&
-              a.params?.action === 'created' &&
-              a.params?.runtimeGroupName === runtimeGroupName
-          );
-          expect(created?.params?.entity).to.equal('RuntimeGroup');
-          expect(created?.params?.action).to.equal('created');
-          expect(created?.params?.runtimeGroupName).to.equal(runtimeGroupName);
-          expect(created?.params?.hostedOrganizations).to.equal(org.name);
-          expect(created?.result).to.equal('success');
-        });
-      });
-    });
-
-    it('records hosting update on organization activity', () => {
-      const runtimeGroupName = `${runtimeGroupId}u`;
-      const hostOrgA = `mof-pups-${runtimeGroupId}`;
-      const hostOrgB = `mof-dogs-${runtimeGroupId}`;
-      const expectedHostedList = [hostOrgA, hostOrgB].sort().join(', ');
-
-      registerHostOrganization(hostOrgA, 'pups').then(
-        ({ apiRes: { status: orgAStatus } }: any) => {
-          expect(orgAStatus).to.be.equal(200);
-
-          registerHostOrganization(hostOrgB, 'dogs').then(
-            ({ apiRes: { status: orgBStatus } }: any) => {
-              expect(orgBStatus).to.be.equal(200);
-
-              cy.setRequestBody({
-                name: runtimeGroupName,
-                hostedOrganizations: [],
-              });
-              cy.callAPI(
-                `ds/api/sdx/v1/organizations/${org.name}/runtime-groups`,
-                'PUT'
-              ).then(({ apiRes: { status: createStatus } }: any) => {
-                expect(createStatus).to.be.equal(200);
-
-                cy.setRequestBody({
-                  name: runtimeGroupName,
-                  hostedOrganizations: [hostOrgA, hostOrgB],
-                });
-                cy.callAPI(
-                  `ds/api/sdx/v1/organizations/${org.name}/runtime-groups`,
-                  'PUT'
-                ).then(({ apiRes: { status: addStatus } }: any) => {
-                  expect(addStatus).to.be.equal(200);
-
-                  cy.callAPI(
-                    `ds/api/sdx/v1/organizations/${org.name}/activity?first=100`,
-                    'GET'
-                  ).then(({ apiRes: { status, body: activities } }: any) => {
-                    expect(status).to.be.equal(200);
-                    const added = activities.find(
-                      (a: any) =>
-                        a.params?.entity === 'RuntimeGroup' &&
-                        a.params?.action === 'updated' &&
-                        a.params?.runtimeGroupName === runtimeGroupName &&
-                        a.params?.hostedOrganizations === expectedHostedList
-                    );
-                    expect(added?.params?.entity).to.equal('RuntimeGroup');
-                    expect(added?.params?.action).to.equal('updated');
-                    expect(added?.params?.hostedOrganizations).to.equal(
-                      expectedHostedList
-                    );
-                    expect(added?.result).to.equal('success');
-
-                    cy.setRequestBody({
-                      name: runtimeGroupName,
-                      hostedOrganizations: [],
-                    });
-                    cy.callAPI(
-                      `ds/api/sdx/v1/organizations/${org.name}/runtime-groups`,
-                      'PUT'
-                    ).then(({ apiRes: { status: clearStatus } }: any) => {
-                      expect(clearStatus).to.be.equal(200);
-
-                      cy.callAPI(
-                        `ds/api/sdx/v1/organizations/${org.name}/activity?first=100`,
-                        'GET'
-                      ).then(
-                        ({ apiRes: { status, body: activities } }: any) => {
-                          expect(status).to.be.equal(200);
-                          const cleared = activities.find(
-                            (a: any) =>
-                              a.params?.entity === 'RuntimeGroup' &&
-                              a.params?.action === 'updated' &&
-                              a.params?.runtimeGroupName === runtimeGroupName &&
-                              a.params?.hostedOrganizations === ''
-                          );
-                          expect(cleared?.params?.entity).to.equal(
-                            'RuntimeGroup'
-                          );
-                          expect(cleared?.params?.action).to.equal('updated');
-                          expect(cleared?.params?.hostedOrganizations).to.equal(
-                            ''
-                          );
-                          expect(cleared?.result).to.equal('success');
-                        }
-                      );
-                    });
-                  });
-                });
-              });
-            }
-          );
-        }
-      );
-    });
-
-    it('records delete on organization activity', () => {
-      const runtimeGroupName = `${runtimeGroupId}d`;
-      // Create first to be able to delete it
-      cy.setRequestBody({
-        name: runtimeGroupName,
-        hostedOrganizations: [org.name],
-      });
-      cy.callAPI(
-        `ds/api/sdx/v1/organizations/${org.name}/runtime-groups`,
-        'PUT'
-      ).then(({ apiRes: { status } }: any) => {
-        expect(status).to.be.equal(200);
-
-        cy.setQueryString({ force: false });
-        cy.callAPI(
-          `ds/api/sdx/v1/organizations/${org.name}/runtime-groups/${runtimeGroupName}`,
-          'DELETE'
-        ).then(({ apiRes: { status: deleteStatus } }: any) => {
-          expect(deleteStatus).to.be.equal(200);
-
-          cy.callAPI(
-            `ds/api/sdx/v1/organizations/${org.name}/activity?first=100`,
-            'GET'
-          ).then(({ apiRes: { status, body: activities } }: any) => {
-            expect(status).to.be.equal(200);
-            const deleted = activities.find(
-              (a: any) =>
-                a.params?.entity === 'RuntimeGroup' &&
-                a.params?.action === 'deleted' &&
-                a.params?.runtimeGroupName === runtimeGroupName
-            );
-            expect(deleted?.params?.entity).to.equal('RuntimeGroup');
-            expect(deleted?.params?.action).to.equal('deleted');
-            expect(deleted?.params?.runtimeGroupName).to.equal(runtimeGroupName);
-            expect(deleted?.result).to.equal('success');
-          });
-        });
-      });
-    });
-  });
-
-  it('keeps v3 organization gateway activity endpoint working', () => {
-    cy.callAPI(
-      `ds/api/v3/organizations/${workingData.org.name}/activity`,
-      'GET'
-    ).then(({ apiRes: { status, body } }: any) => {
-      expect(status).to.be.equal(200)
-      expect(body).to.be.an('array')
     })
   })
 })
